5.1.6 Перенесем функции для рендеринга из `utils.js` в `utils/render.js`
