+++ src/components/task.js
@@ -71,7 +71,7 @@ const createTaskTemplate = (task) => {
 export default class Task extends AbstractComponent {
   constructor(task) {
     super();
-    
+
     this._task = task;
   }
